refactoring: express 서버의 보안관련 모듈의 로직 정리
* 함수들의 역할 재분배
* 불필요한 util 제거
* add todos
