+++ backend/node/express/middleware/authenticate.js
@@ -1,16 +1,22 @@
 import jwt from "jsonwebtoken";
 import config from "../config";
 import {isExistGuest} from "../../DB/queries/guest";
-import {convertPathToEventId} from "../utils";
+import {convertPathToEventCode, isActiveEvent} from "../utils";
 import logger from "../logger.js";
 import CookieKeys from "../CookieKeys.js";
 import {isExistHostOAuthId} from "../../DB/queries/host.js";
+import {getEventByEventCode} from "../../DB/queries/event.js";
 
 const {tokenArgs, routePage} = config;
+
+function isGuestBelongToEvent(guest, event) {
+	return guest.EventId === event.id;
+}
 
 export function guestAuthenticate() {
 	return async function(req, res, next) {
 		const path = req.params.path;
+		const eventCode = convertPathToEventCode(path);
 
 		try {
 			const payload = jwt.verify(
@@ -19,10 +25,15 @@ export function guestAuthenticate() {
 			);
 
 			const guest = await isExistGuest(payload.sub);
-			const eventId = await convertPathToEventId(path);
-			const isGuestBelongToEvent = guest.EventId === eventId;
+			const event = await getEventByEventCode(eventCode);
 
-			if (isGuestBelongToEvent) {
+			// todo better logic flow
+			if (!isActiveEvent(event)) {
+				// todo fix this line of lint
+				throw new Error("이벤트 만료기간이 지났습니다.");
+			}
+
+			if (!isGuestBelongToEvent(guest, event)) {
 				return res.redirect(routePage.guest);
 			}
