+++ backend/node/express/routes/guest.js
@@ -1,17 +1,19 @@
 import express from "express";
 import jwt from "jsonwebtoken";
-import {getTokenExpired} from "../../libs/utils";
 import generateAccessToken from "../authentication/token";
 import config from "../config";
 import {guestAuthenticate} from "../middleware/authenticate";
 import {createGuest, isExistGuest} from "../../DB/queries/guest";
-import {convertPathToEventId} from "../utils";
+import {convertPathToEventCode, getTokenExpired, isActiveEvent} from "../utils";
 import CookieKeys from "../CookieKeys.js";
 import logger from "../logger.js";
 import {AUTHORITY_TYPE_GUEST} from "../../constants/authorityTypes.js";
+import {getEventByEventCode} from "../../DB/queries/event.js";
+
 
 const {routePage, tokenArgs} = config;
 const router = express.Router();
+// todo do somthing
 const cookieExpireTime = 2;
 
 router.get("/", async (req, res) => {
@@ -25,7 +27,6 @@ router.get("/", async (req, res) => {
 
 		if (!guest) {
 			// todo fix this line of lint
-			// noinspection ExceptionCaughtLocallyJS
 			throw Error("Guest is not found");
 		}
 
@@ -41,9 +42,18 @@ router.get("/logout", (req, res) => {
 });
 
 router.get("/:path", guestAuthenticate(), async (req, res) => {
+	const path = req.params.path;
+	const eventCode = convertPathToEventCode(path);
+
 	try {
-		const path = req.params.path;
-		const eventId = await convertPathToEventId(path);
+		const event = await getEventByEventCode(eventCode);
+
+		if (!isActiveEvent(event)) {
+			// todo fix this line of lint
+			throw new Error("이벤트 만료기간이 지났습니다.");
+		}
+
+		const eventId = event.id;
 		const guest = await createGuest(eventId);
 		const accessToken = generateAccessToken(
 			guest.guestSid,
